### 2025-09-16 906ee4a
Uncomment single affiliation
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
