### 2025-09-17 #9001 @danielskatz
@apaolillo - it looks like you still need to address the DOIs as [I mentioned before](https://github.com/openjournals/joss-reviews/issues/9001#issuecomment-3299003481)
