### 2025-09-17 #9001 @apaolillo
Hi @danielskatz thanks for taking care of the initial steps of the process with me. I fixed the build of the paper by attaching my name to my institution through correct indexing. I checked the mentioned references: the first two looks relevant, thus I'm going to include a sentence about them in the related section of the paper. I'm going to check the missing DOIs during the same pass. Stay tuned :)
