### 2025-09-17 f5e8174
JOSS: update related work and bibliography
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
