+++ doc/joss25/paper.bib
@@ -1,21 +1,25 @@
-@misc{nvblox,
+@inproceedings{nvblox,
   title        = {{nvblox: GPU-Accelerated Incremental Signed Distance Field Mapping}},
-  author       = {Alexander Millane and Helen Oleynikova and Emilie Wirbel and Remo Steiner and Vikram Ramasamy and David Tingdahl and Roland Siegwart},
+  author       = {Millane, Alexander and Oleynikova, Helen and Wirbel, Emilie and Steiner, Remo and Ramasamy, Vikram and Tingdahl, David and Siegwart, Roland},
   year         = 2024,
-  url          = {https://arxiv.org/abs/2311.00626},
-  eprint       = {2311.00626},
-  archiveprefix = {arXiv},
-  primaryclass = {cs.RO}
+  booktitle    = {2024 IEEE International Conference on Robotics and Automation (ICRA)},
+  volume       = {},
+  number       = {},
+  pages        = {2698--2705},
+  doi          = {10.1109/ICRA57147.2024.10611532},
+  keywords     = {Surface reconstruction;Navigation;Graphics processing units;Libraries;Path planning;Hardware;Low latency communication}
 }
 
-@misc{robotcore,
+@inproceedings{robotcore,
   title        = {{RobotCore: An Open Architecture for Hardware Acceleration in ROS 2}},
-  author       = {Víctor Mayoral-Vilches and Sabrina M. Neuman and Brian Plancher and Vijay Janapa Reddi},
-  year         = 2023,
-  url          = {https://arxiv.org/abs/2205.03929},
-  eprint       = {2205.03929},
-  archiveprefix = {arXiv},
-  primaryclass = {cs.RO}
+  author       = {Mayoral-Vilches, Víctor and Neuman, Sabrina M. and Plancher, Brian and Reddi, Vijay Janapa},
+  year         = 2022,
+  booktitle    = {2022 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  volume       = {},
+  number       = {},
+  pages        = {9692--9699},
+  doi          = {10.1109/IROS47612.2022.9982082},
+  keywords     = {Computer architecture;Benchmark testing;Software;Real-time systems;Workstations;Time factors;Robots}
 }
 
 @inproceedings{tani2020reproducible,
@@ -186,4 +190,30 @@
   eprint       = {1305.4584},
   archiveprefix = {arXiv},
   primaryclass = {cs.PL}
+}
+
+@article{ritchie2020caliban,
+  title        = {{Caliban: Docker-based job manager for reproducible workflows}},
+  author       = {Ritchie, Sam and Slone, Ambrose and Ramasesh, Vinay},
+  year         = 2020,
+  journal      = {Journal of Open Source Software},
+  publisher    = {The Open Journal},
+  volume       = 5,
+  number       = 53,
+  pages        = 2403,
+  doi          = {10.21105/joss.02403},
+  url          = {https://doi.org/10.21105/joss.02403}
+}
+
+@article{viereck2019x11docker,
+  title        = {{x11docker: Run GUI applications in Docker containers}},
+  author       = {Viereck, Martin},
+  year         = 2019,
+  journal      = {Journal of Open Source Software},
+  publisher    = {The Open Journal},
+  volume       = 4,
+  number       = 37,
+  pages        = 1349,
+  doi          = {10.21105/joss.01349},
+  url          = {https://doi.org/10.21105/joss.01349}
 }
