### 2025-09-18 #9001 @danielskatz
👋  @DrDaveD & @gvallee & @cnewburn, would one or more of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
