### 2025-09-19 #9001 @cnewburn
Dan

I can't fit this in right now but I can ask other senior people in the
containers community if you wish.

On Thu, Sep 18, 2025 at 8:11 AM Daniel S. Katz ***@***.***>
wrote:

> *danielskatz* left a comment (openjournals/joss-reviews#9001)
> <https://github.com/openjournals/joss-reviews/issues/9001#issuecomment-3307375158>
>
> 👋 @DrDaveD <https://github.com/DrDaveD> & @gvallee
> <https://github.com/gvallee> & @cnewburn <https://github.com/cnewburn>,
> would one or more of you be willing to review this submission for JOSS? We
> carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
