### 2025-09-22 #9059 @editorialbot
[REVIEW]: pythainer: composable and reusable Docker builders and runners for reproducible research
**Submitting author:** @apaolillo (<a href="http://orcid.org/0000-0001-6608-6562">Antonio Paolillo</a>)
**Repository:** https://github.com/apaolillo/pythainer
**Branch with paper.md** (empty if default branch): main
**Version:** v0.0.6
