### 2025-09-26 #9059 @samfrm
GUI support is broken. Running the example on my Kubuntu laptop shows Docker being launched with an environment variable whose name is my host’s `Xauthority` path (`--env=/run/user/1000/xauth…=/root/.Xauthority`), and there’s no mount for that file. Because of that, `xeyes` (and anything else using X) exits immediately. Probably the fix is to export `XAUTHORITY=<host path>` and bind-mount that file into the container. see `src/pythainer/examples/runners/__init__.py:38`. While you’re in there, you might also drop the unconditional `--tty` flag so scripted runs don’t trip the `“input device is not a TTY”` warning.
