### 2025-09-28 #9059 @apaolillo
Dear @samfrm . Thanks for flagging. I installed a KUbuntu version and the GUI support worked (see screenshot).

May I ask you the pythainer script you used to trigger the issue? My MWE is the following:
```python
from pythainer.examples.builders import get_user_builder
from pythainer.examples.runners import gpu_runner, gui_runner
from pythainer.runners import ConcreteDockerRunner

image = "devtools"
b = get_user_builder(image_name=image, base_ubuntu_image="ubuntu:24.04")
b.root()
b.add_packages(["x11-apps"])
b.user()
b.workdir("/home/user")
b.build()

r = b.get_runner()
r |= gui_runner()
r.run()
```

Feel free to create an issue directly on [pythainer](https://github.com/apaolillo/pythainer) repo to discuss further.
