### 2025-10-20 fe6a246
revamp paths
Signed-off-by: Bogaert Aaron <aaron.bogaert@hotmail.com>
