### 2025-10-20 41e123f
documentation and bugfixes
Signed-off-by: Bogaert Aaron <aaron.bogaert@hotmail.com>
