### 2025-10-22 27ef879
Delete user by UID instead of UID:GID
When the host does not match the guest way of handling users it will
break. Removing users by UID is more robust

Signed-off-by: Esteban Aguililla Klein <esteban.aguililla.klein.pro@outlook.com>
