### 2025-10-28 #9059 @samfrm
Thank you @apaolillo  for submitting pythainer. I reviewed the repo, docs, tests/CI, and the JOSS paper.

**Summary**

- pythainer lets you build Docker images by composing small Python “builders,” and run them with reusable “runners” (GPU, GUI, devices, volumes).
- It stays Docker‑native: renders standard Dockerfiles, builds with Docker, and runs containers with clear flags.
- The CLI supports both scaffolding a starter script and running directly, which makes it easy to adopt.

**Checklist Summary**

- Functionality: Works as described; examples are practical (QEMU, LLVM/MLIR).
- Documentation: README includes motivation, quick start, examples, and clear project layout.
- Installation: Straightforward via PyPI or source; Python 3.10+ noted.
- Examples: Runnable, relevant, and easy to adapt.
- Tests: Unit tests for deterministic Dockerfile rendering and CLI; optional Docker integration test.
- CI: GitHub Actions run tests and format/lint checks on multiple Python versions.
- License: MIT license present.
- Community Guidelines: CONTRIBUTING and MAINTAINERS present.
- Paper: Clear statement of need, scope, functionality, and related work; appropriate for JOSS.

**Paper**

- Statement of need is clear for research groups using Docker.
- Functionality and audience are well explained.
- Related work positions pythainer fairly.
- Writing is concise and readable.

**Verdict**

- I found no blockers. The software is installable, documented, and tested; the paper is in scope for JOSS.
- Recommendation: Accept.

**Non‑Blocking Suggestions**

- Add a simple CITATION.cff.
- Create an archived release (e.g., Zenodo) and include the DOI in the JOSS metadata.
