### 2025-10-28 #9059 @samfrm
> > **Non‑Blocking Suggestions**
> >
> > * Add a simple CITATION.cff.
> > * Create an archived release (e.g., Zenodo) and include the DOI in the JOSS metadata.
>
> [@danielskatz](https://github.com/danielskatz) should I address this prior to (potential) publication or is it something independent?

An archived release is something I highly recommend to have (And I guess it is required before JOSS publication? @danielskatz )
