### 2025-10-29 apaolillo/pythainer#40 @rcannood
[JOSS] Review of Documentation
Hi @apaolillo ! I'm reviewing your submission to JOSS as part of openjournals/joss-reviews#9059!

---

While the documentation is mostly satisfactory, I feel like it would benefit from separating out the different sections and turning it into something that gives a bit more of an overview, e.g. a readthedocs.

---

>  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The README documents this, though the section "Why Pythainer?" is too detailed and too on the nose generated by an LLM that as a potential user I would not like to actually read it -- perhaps the section "In short" should be moved up in the readme, since this is actually what I would like to read?

---

>  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

There are sufficient examples added to the README, but essential documentation that I would need to actually get started with pythainer is:

* Document how different parts of recipes result in which Dockerfiles, and how a user would be able to view the Dockerfile before building it? It definitely is a benefit that pythainer is able to generate dockerfiles for you, but I still want to know what is happening behind the screens so I know how to use the available builders and how this translates into which Dockerfile.

* The different builders and runners need to be documented better; right now it just points to the codebase:
