### 2025-12-11 9050eda
Add RUN --mount support and ssh forwarding
Introduce the RunMount API for expressing BuildKit RUN --mount clauses
and extend PartialDockerBuilder.run() to accept mount lists while
tracking ssh requirement.

DockerBuilder now forwards the ssh agent when ssh mounts are used.

Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
