### 2025-12-11 c0720a7
Fix logic error
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
