### 2025-12-12 c8864bc
Add libslirp as a QEMU dependency (#43)
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
