+++ examples/copy/copy_container.py
@@ -1,14 +1,16 @@
 import os
 from pathlib import Path
+
 from pythainer.examples.builders import get_user_builder
+
 
 def main():
     builder = get_user_builder(
-            image_name="test",
-            base_ubuntu_image="ubuntu:24.04",
-            user_name="jeffry",
-            packages=["python3", "python3-pip", "python3-venv", "python3-dev", "zsh"]
-        )
+        image_name="test",
+        base_ubuntu_image="ubuntu:24.04",
+        user_name="jeffry",
+        packages=["python3", "python3-pip", "python3-venv", "python3-dev", "zsh"],
+    )
 
     p = Path("/home/${USER_NAME}/workspace/libraries")
 
