+++ examples/intel_vtune.py
@@ -3,19 +3,16 @@
 from pythainer.examples.builders import get_user_gui_builder, vTune_builder
 from pythainer.examples.runners import gui_runner
 from pythainer.runners import ConcreteDockerRunner
-from pythainer.sysutils import shell_out
+
 image_name = "pythainertest"
 builder = get_user_gui_builder(image_name=image_name, base_ubuntu_image="ubuntu:22.04")
 builder.root()
 builder.add_packages(packages=["vim", "git", "tmux"])
 builder.user("${USER_NAME}")
 builder |= vTune_builder()
-# builder |= ssh_builder()
-runner = ConcreteDockerRunner(image=image_name,volumes={"/usr/src/":"/usr/src/"})
+runner = ConcreteDockerRunner(image=image_name, volumes={"/usr/src/": "/usr/src/"})
 guiRun = gui_runner()
 builder.build()
 runner |= guiRun
 
 runner.run()
-
-# sudo ./build-driver --c-compiler=/usr/bin/gcc-12 --kernel-src-dir=/usr/src/linux-headers-6.8.0-52-generic/
