### 2025-12-12 02f907d
split string
Signed-off-by: Bogaert Aaron <aaron.bogaert@hotmail.com>
