+++ src/pythainer/examples/builders/__init__.py
@@ -280,7 +280,8 @@ def vTune_builder(
             "wget https://apt.repos.intel.com/intel-gpg-keys/GPG-PUB-KEY-INTEL-SW-PRODUCTS.PUB",
             "sudo apt-key add GPG-PUB-KEY-INTEL-SW-PRODUCTS.PUB",
             "rm GPG-PUB-KEY-INTEL-SW-PRODUCTS.PUB",
-            'echo "deb https://apt.repos.intel.com/oneapi all main" | sudo tee /etc/apt/sources.list.d/oneAPI.list',
+            ('echo "deb https://apt.repos.intel.com/oneapi all main"'
+             ' | sudo tee /etc/apt/sources.list.d/oneAPI.list'),
         ]
     )
 
