+++ examples/intel_vtune.py
@@ -14,7 +14,7 @@ To run the GUI version of the tool in the container use:
 /opt/intel/oneapi/vtune/latest/bin64/vtune-gui
 """
 
-from pythainer.examples.builders import get_user_gui_builder, vTune_builder
+from pythainer.examples.builders import get_user_gui_builder, vtune_builder
 from pythainer.examples.runners import gui_runner
 from pythainer.runners import ConcreteDockerRunner
 
@@ -28,7 +28,7 @@ def main():
     builder.root()
     builder.add_packages(packages=["vim", "git", "tmux"])
     builder.user("${USER_NAME}")
-    builder |= vTune_builder()
+    builder |= vtune_builder()
     runner = ConcreteDockerRunner(image=image_name, volumes={"/usr/src/": "/usr/src/"})
     gui_run = gui_runner()
     builder.build()
