### 2025-12-12 5e7bd6e
Fix: allow Python >= 3.10 in install_venv.sh (#45)
Signed-off-by: Mats Van Molle <mats.vanmolle@icloud.com>
