+++ src/pythainer/builders/commands/run/mounts.py
@@ -1,0 +1,354 @@
+# Copyright (C) 2025 Antonio Paolillo. All rights reserved.
+# SPDX-License-Identifier: MIT
+"""
+Utilities for expressing `RUN --mount` clauses in a Dockerfile.
+
+This module provides the `RunMount` dataclass, which models a single mount
+specification to be used with Docker BuildKit's extended syntax:
+
+    RUN --mount=type=...,option=value,... <command>
+
+Each mount corresponds to a named mount type (`bind`, `cache`, `tmpfs`,
+`secret`, `ssh`) and a mapping of mount-specific options.
+
+The convenience class methods (`cache`, `bind`, `tmpfs`, `secret`, `ssh`) provide
+typed constructors that mirror Docker's documented mount options and generate a
+`RunMount` instance ready for serialization via `to_flag()`.
+
+The output of `to_flag()` is intended to be consumed by the Dockerfile builder,
+which concatenates several mounts and produces a full instruction like:
+
+    RUN --mount=type=cache,target=/root/.cache/go-build go build ...
+
+The `RunMount` objects themselves are purely declarative and contain no
+Docker-specific logic beyond formatting.
+"""
+
+from dataclasses import dataclass, field
+from typing import Dict, Union
+
+Scalar = Union[str, int, bool]
+
+
+@dataclass
+class RunMount:
+    """
+    Represents a single Docker BuildKit mount clause for a `RUN` instruction.
+
+    A `RunMount` encapsulates both the mount type (e.g., `"cache"`, `"bind"`,
+    `"secret"`) and the dictionary of mount-specific options. The instance can
+    be converted to a valid Docker CLI flag using `to_flag()`, producing:
+
+        --mount=type=<mount_type>,key=value,key2=value2,...
+
+    The class also provides convenience constructors (for example, `cache()`,
+    `secret()`, `ssh()`) that populate the correct option fields for the
+    respective mount types and mirror the documented BuildKit semantics.
+
+    Attributes:
+        mount_type (str):
+            The Docker mount type (for example, `"cache"`, `"bind"`, `"tmpfs"`,
+            `"secret"`, `"ssh"`).
+
+        options (dict[str, Scalar]):
+            Dictionary of mount options. Keys correspond to BuildKit option
+            names such as `target`, `sharing`, `mode`, `uid`, etc. Values may be
+            strings, integers, or booleans. Boolean values follow BuildKit
+            semantics: a key is emitted without a value if it is `True`, and
+            omitted entirely if `False`.
+    """
+
+    mount_type: str
+    options: Dict[str, Scalar] = field(default_factory=dict)
+
+    def to_flag(self) -> str:
+        """
+        Convert this mount specification into a Docker `--mount=` flag.
+
+        Returns:
+            str: The full `--mount=...` expression. Example:
+
+                "--mount=type=cache,target=/root/.cache/go-build,sharing=locked"
+
+        Notes:
+            * Boolean options encode to flags without a value (e.g. `rw`, `readonly`).
+            * Options set to `False` are omitted entirely.
+        """
+        parts: list[str] = [f"type={self.mount_type}"]
+        for key, value in self.options.items():
+            if isinstance(value, bool):
+                # BuildKit behavior: boolean=True means emit bare key.
+                if value:
+                    parts.append(key)
+            else:
+                parts.append(f"{key}={value}")
+        return "--mount=" + ",".join(parts)
+
+    # Convenience constructors
+
+    @classmethod
+    def cache(
+        cls,
+        target: str,
+        *,
+        id_: str | None = None,
+        ro: bool | None = None,
+        readonly: bool | None = None,
+        sharing: str | None = None,
+        source: str | None = None,
+        from_: str | None = None,
[…]
