+++ src/pythainer/builders/commands/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (C) 2025 Antonio Paolillo. All rights reserved.
+# SPDX-License-Identifier: MIT
