### 2025-12-13 4efca37
Avoid 0 UID & GID being passed to docker build
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
