+++ src/pythainer/builders/__init__.py
@@ -575,8 +575,8 @@ class UbuntuDockerBuilder(DockerBuilder):
             username (str): The username of the new user.
         """
         self.arg(name="USER_NAME", value=username)
-        self.arg(name="UID", value="1000") # default value if UID not provided
-        self.arg(name="GID", value="1000") # default value if GID not provided
+        self.arg(name="UID", value="1000")  # default value if UID not provided
+        self.arg(name="GID", value="1000")  # default value if GID not provided
         self.remove_group_if_gid_exists(gid="${GID}")
         self.remove_user_if_uid_exists(uid="${UID}", gid="${GID}")
         self.run(command="groupadd -g ${GID} ${USER_NAME}")
