### 2025-12-13 0d27b0f
Avoid 0 UID & GID being passed to docker build (#46)
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
