### 2025-12-14 ef9e9d8
Address reviewer comments on paper
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
