### 2025-12-14 edcd330
Address reviewer comments on paper
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
