### 2025-12-15 b3b6d4b
Delete user by UID instead of UID:GID (#42)
When the host does not match the guest way of handling users it will
break.  Removing users by UID is more robust. By default, we take the
host UID and GID, if both do not match it will fail at deleting the
default user therefore, failing at building the pythainer.

Signed-off-by: Esteban Aguililla Klein <esteban.aguililla.klein.pro@outlook.com>
