### 2025-12-15 2956488
Add minimum support for VTune profiler (#14)
This introduces a minimal, working integration of Intel VTune into the
Docker container.

The sampling driver is not included yet due to a GCC toolchain conflict,
which means a large part of VTune's analysis features are still
unavailable.

To run the GUI version:
  /opt/intel/oneapi/vtune/latest/bin64/vtune-gui

Signed-off-by: Aaron Bogaert <aaron.bogaert@hotmail.com>
