### 2025-12-15 2250f9f
Improve README.md according to JOSS review
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
