### 2025-12-15 apaolillo/pythainer#40 @apaolillo
Hi @rcannood ,

Thanks again for the detailed and constructive review. I really appreciated the "new user" perspective.

I've now opened a PR that implements the documentation changes discussed in your comment (README reorganization for first-time users, a short "How it works / Dockerfile generation" section, and an excerpt showing what the generated Dockerfile looks like + where to inspect it). This should make it easier to understand how builders translate into Dockerfile blocks, without needing to read the implementation.

The only item not included in this PR is the "proper documentation site" (ReadTheDocs-style) restructuring, since that feels like a larger reorganization best handled separately from the JOSS review timeline. If you agree, I propose to split that into a dedicated GitHub issue and address it independently (still on the roadmap, just decoupled from the paper review).

Could you please have a look at the PR #48 and let me know whether it addresses the core usability/documentation concerns you raised?

Best,
Antonio
