### 2025-12-16 7225a6a
style changes
Signed-off-by: Bogaert Aaron <aaron.bogaert@hotmail.com>
