### 2025-12-16 992aeb1
removed unneeded pylint
Signed-off-by: Bogaert Aaron <aaron.bogaert@hotmail.com>
