### 2025-12-16 f0cf9d9
fixed bug that occurs when no copy is pressent on first use of pythainer
Signed-off-by: Bogaert Aaron <aaron.bogaert@hotmail.com>
