### 2025-12-16 0d3edc2
Improve README.md according to JOSS review (#48)
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
