### 2025-12-16 9de2668
Make Docker COPY usable via builder.copy() (#28)
Introduce a proper CopyDockerBuildCommand that stages host files and
directories into the Docker build context and emits a COPY --chown
directive in the generated Dockerfile.

Update PartialDockerBuilder.copy() to use the new command instead of
injecting raw strings, and extend the build flow to include staged data
in the temporary build directory.

Add a small example demonstrating how to copy files from the host into
an image at build time.

Signed-off-by: Bogaert Aaron <aaron.bogaert@hotmail.com>
