+++ src/pythainer/builders/cmds.py
@@ -9,7 +9,11 @@ The module facilitates the dynamic generation of Dockerfile content based on dif
 and package managers.
 """
 
+import shutil
+from pathlib import Path
 from typing import List
+
+from pythainer.sysutils import mkdir
 
 
 class DockerBuildCommand:
@@ -67,6 +71,58 @@ class StrDockerBuildCommand(DockerBuildCommand):
             str: The command string.
         """
         return str(self._str)
+
+
+class CopyDockerBuildCommand(DockerBuildCommand):
+    """
+    Represents the command string to copy data from the host system to
+    the docker container at build time.
+    """
+
+    def __init__(self, source_path: Path, destination_path: Path) -> None:
+        """
+        Initializes the CopyDockerBuildCommand with a a source and destination path.
+
+        Parameters:
+            source_path (Path): Path of folder or file to copy to container
+            destination_path (Path): Path to copy the file or folder to
+        """
+        super().__init__()
+        self._source_path = source_path.resolve()
+        self._destination_path = destination_path
+
+    def get_str_for_dockerfile(
+        self,
+        *args,
+        **kwargs,
+    ) -> str:
+        """
+        Generates a Dockerfile string to move files and folders.
+
+        Returns:
+            str: A Dockerfile command string for moving files and folders.
+        """
+
+        data_path = Path("/tmp/pythainer/docker/data")
+        resulting_path = data_path / self._source_path.relative_to("/")
+        relative_path = Path("data") / self._source_path.relative_to("/")
+        mkdir(data_path)
+
+        print(data_path)
+
+        if self._source_path.is_file():
+            mkdir(resulting_path.parent)
+            shutil.copyfile(self._source_path, resulting_path)
+        elif self._source_path.is_dir():
+            shutil.copytree(self._source_path, resulting_path, dirs_exist_ok=True)
+        else:
+            raise FileExistsError(
+                f"{self._source_path} is not a valid target to copy into the docker container"
+            )
+
+        cmd = f"COPY --chown=${{USER_NAME}} {relative_path} {self._destination_path}"
+
+        return cmd
 
 
 class AddPkgDockerBuildCommand(DockerBuildCommand):
