### 2025-12-16 1649606
Introduce explicit build context for COPY
Replace ad-hoc filesystem staging with a first-class BuildContext that
tracks host paths and materializes a proper Docker build context.

Rework builder.copy() to register sources in the context, support
multi-source COPY, and forward --chown/--chmod flags explicitly.
Make CopyDockerBuildCommand pure by restricting it to context-relative
paths and Dockerfile string generation only.

Update the build flow to construct and use the generated context
directory, and clean up the COPY example accordingly.

Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
