+++ src/pythainer/builders/__init__.py
@@ -7,7 +7,6 @@ handling commands like package installation, environment variable setting, and u
 tailored specifically for Docker environments.
 """
 import os
-import shutil
 import tempfile
 from pathlib import Path
 from typing import Dict, Iterable, List
@@ -19,6 +18,7 @@ from pythainer.builders.cmds import (
     StrDockerBuildCommand,
 )
 from pythainer.builders.commands.run.mounts import RunMount
+from pythainer.builders.contexts import BuildContext
 from pythainer.runners import ConcreteDockerRunner
 from pythainer.sysutils import (
     PathType,
@@ -59,11 +59,15 @@ class PartialDockerBuilder:
     merged.
     """
 
-    def __init__(self) -> None:
+    def __init__(
+        self,
+        context_root: PathType | None = None,
+    ) -> None:
         """
         Initializes a PartialDockerBuilder with an empty list of build commands.
         """
         self._build_commands: List[DockerBuildCommand] = []
+        self._context: BuildContext = BuildContext(context_root=context_root)
         self._needs_ssh: bool = False
 
     def __or__(self, other: "PartialDockerBuilder") -> "PartialDockerBuilder":
@@ -104,6 +108,7 @@ class PartialDockerBuilder:
         """
         # pylint: disable=protected-access
         self._build_commands.extend(other._build_commands)
+        self._context.extend(other._context)
         self._needs_ssh = self._needs_ssh or other._needs_ssh
 
     def space(self) -> None:
@@ -248,16 +253,40 @@ class PartialDockerBuilder:
         """
         self._build_commands.append(AddPkgDockerBuildCommand(packages=packages))
 
-    def copy(self, source_path: Path, destination_path: Path) -> None:
-        """
-        Copies a file to the docker container
+    def copy(
+        self,
+        source: Path | list[Path],
+        destination: Path,
+        chown: str | None = None,
+        chmod: str | None = None,
+    ) -> None:
+        """Add a Dockerfile COPY instruction.
 
-        Parameters:
-            source_path (Path): The file or folder to copy to the container.
-            destination_path (Path): The location to place the
-                file or folder within the Docker container.
+        Args:
+            source: A single host path or an iterable of host paths (file or dir).
+            destination: Container destination path.
+            chown: Optional ownership, forwarded to Dockerfile `COPY --chown=...`.
+            chmod: Optional permissions, forwarded to Dockerfile `COPY --chmod=...`.
+
+        Raises:
+            ValueError: If source list is empty or destination is invalid for multi-source copy.
         """
-        self._build_commands.append(CopyDockerBuildCommand(source_path, destination_path))
+        sources = (source,) if isinstance(source, Path) else tuple(source)
+        if not sources:
+            raise ValueError("copy(): at least one source path is required")
+
+        ctx_paths = [
+            self._context.add_context_entry(host_path=source_path) for source_path in sources
+        ]
+
+        self._build_commands.append(
+            CopyDockerBuildCommand(
+                sources=ctx_paths,
+                destination=destination,
+                chown=chown,
+                chmod=chmod,
+            )
+        )
 
 
 class DockerBuilder(PartialDockerBuilder):
@@ -429,14 +458,12 @@ class DockerBuilder(PartialDockerBuilder):
             dockerfile_savepath (PathType): Optional path to save the Dockerfile used for the build.
             docker_context (PathType): Optional path to save the Docker context used for the build.
         """
-
         main_dir = Path("/tmp/pythainer/docker/")
         mkdir(main_dir)
         with tempfile.TemporaryDirectory(
             prefix="/tmp/pythainer/docker/docker-build-",
             dir=main_dir,
         ) as temp_dir:
-
             temp_path = Path(temp_dir)
[…]
