+++ examples/copy/copy_container.py
@@ -3,16 +3,28 @@
 # SPDX-License-Identifier: MIT
 
 """
-Example of using the docker COPY directive.
+This example demonstrates how to copy files from the host filesystem
+into a Docker image at build time using ``builder.copy(...)``.
 
-The builder.copy takes two path arguments
-The first a path to the file on the host system you want to copy
-and second a path inside of the the docker image where you want these files to be copied over to.
+Overview
+--------
+- `builder.copy()` registers one or more host-side paths to be included
+  in the Docker build context.
+- These paths are then copied into the image using a Dockerfile
+  `COPY` instruction.
+- The destination path refers to a location *inside the image*.
+- Source paths may be absolute or relative on the host system.
 
-These Paths can be absolute as well as relative.
+Notes
+-----
+- All staging of host files into the build context is handled by the
+  builder; Dockerfile instructions only reference context-relative paths.
+- Ownership and permissions can be controlled using `chown` and
+  `chmod` arguments, which are forwarded to the Dockerfile `COPY`
+  instruction.
 """
 
-import os
+
 from pathlib import Path
 
 from pythainer.examples.builders import get_user_builder
@@ -23,16 +35,22 @@ def main():
     Main entry point: builds and runs COPY example.
     """
     builder = get_user_builder(
-        image_name="test",
+        image_name="pythnrcopy",
         base_ubuntu_image="ubuntu:24.04",
         user_name="jeffry",
         packages=["python3", "python3-pip", "python3-venv", "python3-dev", "zsh"],
     )
 
-    p = Path("/home/${USER_NAME}/workspace/libraries")
+    p = Path("/home/${USER_NAME}/workspace/test")
     builder.run("mkdir {p}")
 
-    builder.copy(Path(os.path.dirname(__file__)) / "file_to_coppy.txt", p / "resulting_file.txt")
+    data_dir = Path(__file__).parent
+    file_to_copy = data_dir / "file_to_coppy.txt"
+    builder.copy(
+        source=file_to_copy,
+        destination=p / "resulting_file.txt",
+        chown="${USER_NAME}:${USER_NAME}",
+    )
 
     builder.workdir(p)
     builder.build()
