+++ src/pythainer/builders/contexts.py
@@ -1,0 +1,161 @@
+# Copyright (C) 2025 Antonio Paolillo. All rights reserved.
+# SPDX-License-Identifier: MIT
+"""
+Build context management for Docker image construction.
+
+This module defines the `BuildContext`, which is responsible for
+constructing a Docker build context from a set of host-side files and
+directories. The build context is a *materialized directory tree* that
+Docker can consume during `docker build`.
+
+This module is *host-aware* and performs filesystem operations. Dockerfile AST
+nodes will further reference context-relative paths. Each context entry maps a
+host path to a unique, relative path inside the build context. Context paths
+are validated to prevent path traversal (`..`). Conflicting context entries
+are detected and rejected.
+
+The build context is populated explicitly via `add_context_entry` and can be
+merged with other contexts via `extend`. Once fully defined, the context can be
+materialized on disk using `build` (from the DockerBuilder class).
+"""
+
+import shutil
+from pathlib import Path
+
+from pythainer.sysutils import PathType
+
+
+class BuildContext:
+    """
+    Represents a Docker build context under construction.
+
+    A `BuildContext` maintains a mapping from *context-relative paths*
+    (paths as seen by the Dockerfile) to *host filesystem paths*. It is
+    responsible for materializing this mapping into a directory tree that
+    can be passed to `docker build`.
+
+    Attributes:
+        _context_root:
+            Optional root directory used to compute context-relative paths.
+            If provided, all added host paths must be located under this root.
+        _context_entries:
+            Mapping from context-relative paths to host filesystem paths.
+    """
+
+    def __init__(self, context_root: PathType | None = None) -> None:
+        """
+        Initialize an empty build context.
+
+        Args:
+            context_root (PathType, optional):
+                Optional directory that defines the root of the build context.
+                If set, added host paths are made relative to this root when
+                computing their context paths. If `None`, only the basename
+                of each host path is used.
+        """
+        self._context_root = Path(context_root) if context_root else None
+        self._context_entries: dict[Path, Path] = {}  # ctx_path -> host_path
+
+    def extend(self, other: "BuildContext") -> None:
+        """
+        Merge another build context into this one.
+
+        All context entries from `other` are added to this context. If a
+        context-relative path already exists in both contexts and refers to
+        different host paths, the merge fails.
+
+        Args:
+            other: Another `BuildContext` to merge.
+
+        Raises:
+            ValueError: If conflicting context entries are detected.
+        """
+        # pylint: disable=protected-access
+        collision_paths = self._context_entries.keys() & other._context_entries.keys()
+        diff_collision_paths = {
+            str(p) for p in collision_paths if self._context_entries[p] != other._context_entries[p]
+        }
+        if diff_collision_paths:
+            raise ValueError(
+                f"Duplicate context entries detected before merging: {diff_collision_paths}"
+            )
+
+        self._context_entries.update(other._context_entries)
+
+    def add_context_entry(self, host_path: PathType) -> Path:
+        """
+        Register a host file or directory in the build context.
+
+        The host path is mapped to a context-relative path. If `context_root`
+        was provided at construction time, the context path is computed as the
+        path of `host_path` relative to `context_root`. Otherwise, the basename
+        of `host_path` is used.
+
+        Args:
+            host_path: Path to a file or directory on the host filesystem.
+
+        Returns:
+            The context-relative path under which the entry will appear in the
+            build context.
+
+        Raises:
+            ValueError: If a different host path is already registered under
+                the same context-relative path.
+        """
+        host_path = Path(host_path)
+
+        if self._context_root is not None:
[…]
