### 2025-12-22 74dc65b
Add examples testing variants of the feature
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
