+++ examples/copy/copy_container.py
@@ -1,39 +1,68 @@
 #!/usr/bin/env python3
-# Copyright (C) 2025 Aaron Bogaert. All rights reserved.
+# Copyright (C) 2025 Vrije Universiteit Brussel. All rights reserved.
 # SPDX-License-Identifier: MIT
 
 """
-This example demonstrates how to copy files from the host filesystem
-into a Docker image at build time using ``builder.copy(...)``.
+Pythainer example: COPY with PartialDockerBuilder and build contexts.
 
-Overview
---------
-- `builder.copy()` registers one or more host-side paths to be included
-  in the Docker build context.
-- These paths are then copied into the image using a Dockerfile
-  `COPY` instruction.
-- The destination path refers to a location *inside the image*.
-- Source paths may be absolute or relative on the host system.
+This example demonstrates how `PartialDockerBuilder` handles Docker COPY
+instructions when multiple host files share the same basename, and how
+providing an explicit build context (`context_root`) resolves ambiguities.
 
-Notes
------
-- All staging of host files into the build context is handled by the
-  builder; Dockerfile instructions only reference context-relative paths.
-- Ownership and permissions can be controlled using `chown` and
-  `chmod` arguments, which are forwarded to the Dockerfile `COPY`
-  instruction.
+Key concepts illustrated
+------------------------
+1. **PartialDockerBuilder**:
+   A lightweight builder that records Dockerfile fragments (e.g., COPY
+   instructions) without immediately attaching them to a full image build.
+
+2. **Build context ambiguity**:
+   Docker COPY requires a single build context root. When two host paths with
+   identical filenames but different directories are copied without an explicit
+   context root, ambiguity arises.
+
+3. **context_root**:
+   By explicitly setting `context_root`, Pythainer can safely map multiple host
+   files with overlapping names into the image without collisions.
+
+What this example does
+----------------------
+1. Creates a base image using `get_user_builder`.
+2. Creates a target directory inside the image filesystem.
+3. Creates a temporary directory tree on the host with two files:
+   - `/tmp/pythainer/test/file.txt`
+   - `/tmp/pythainer/test/nested/file.txt`
+4. Demonstrates that copying both files *without* a context root raises an error.
+5. Demonstrates that copying the same files *with* a context root succeeds.
+6. Merges the valid partial builder into the main builder.
+7. Builds and runs the resulting image.
+
+Why this matters
+----------------
+This pattern is important when:
+- You want to assemble Dockerfile fragments incrementally.
+- You copy many files from the host with overlapping names.
+- You want explicit, reproducible control over Docker build contexts.
 """
-
 
 from pathlib import Path
 
+from pythainer.builders import PartialDockerBuilder
 from pythainer.examples.builders import get_user_builder
 
 
-def main():
+def main() -> None:
     """
-    Main entry point: builds and runs COPY example.
+    Build and run an example image demonstrating PartialDockerBuilder COPY rules.
+
+    The script constructs two host files with identical names in different
+    directories and shows how Pythainer enforces explicit context roots to avoid
+    ambiguous Docker COPY semantics.
+
+    Raises:
+        AssertionError: If the expected COPY ambiguity error is not raised.
     """
+    # Base image builder with user and common packages.
+    # pylint: disable=duplicate-code
     builder = get_user_builder(
         image_name="pythnrcopy",
         base_ubuntu_image="ubuntu:24.04",
@@ -41,19 +70,75 @@ def main():
         packages=["python3", "python3-pip", "python3-venv", "python3-dev", "zsh"],
     )
 
-    p = Path("/home/${USER_NAME}/workspace/test")
-    builder.run("mkdir {p}")
[…]
