### 2025-12-22 cbbe496
Add documentation for context_root path
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
