### 2026-01-07 38f4623
Add source for JOSS'25 paper (#38)
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
