+++ doc/joss25/paper.md
@@ -1,0 +1,172 @@
+---
+title: "pythainer: composable and reusable Docker builders and runners for reproducible research"
+tags:
+  - Python
+  - Docker
+  - reproducibility
+  - software engineering
+  - research tooling
+  - system software
+authors:
+  - name: Antonio Paolillo
+    orcid: 0000-0001-6608-6562
+    affiliation: 1
+affiliations:
+  - name: Software Languages Lab, Vrije Universiteit Brussel (VUB), Belgium
+    index: 1
+    ror: 006e5kg04
+date: 16 September 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+Software experiments today often depend on complex Linux environments that
+combine several toolchains, devices, and graphical interfaces. Many research
+projects [@nvblox; @robotcore], for instance, need to compose ROS 2
+[@macenski2022ros2] with CUDA [@nvidiaCudaPg13], require non-root users, provide
+GPU and GUI access, and must be reproducible across time and machines. Docker
+[@docker] is a widely adopted substrate for packaging and running such
+environments, and is commonly used to improve reproducibility in research
+software [@tani2020reproducible]. However, writing and maintaining Dockerfiles
+and project-specific `docker run` scripts becomes a burden as requirements grow.
+
+`pythainer` raises the level of abstraction while remaining Docker-native. It
+lets users describe images as small, testable Python *builders* that can be
+composed (e.g., ROS 2 + CUDA) and executed with reusable *runners* that capture
+runtime policy (GPU, GUI, users, mounts). `pythainer` renders deterministic
+Dockerfiles, builds standard images, and centralizes run configuration, hence
+improving reuse and reducing duplication across repositories.
+
+# Statement of need
+
+Plain Dockerfiles are intentionally minimal: they offer sequential shell steps
+but no first-class functions, loops, or composition. This is adequate for simple
+images, yet it complicates reuse in research settings where environments must be
+combined and parameterized. In particular, merging two existing images (e.g.,
+community ROS 2 and NVIDIA CUDA) is not first-class; multi-stage builds help trim
+artifacts but require intimate knowledge of which files, environment variables,
+and paths must be copied and preserved. On the runtime side, real projects often
+need non-root users, persistent volumes, access to GPUs and GUIs (X11/Wayland),
+and device mappings. These concerns are typically maintained as long shell
+scripts that are copy-pasted and diverge across projects.
+
+The primary target audience of `pythainer` is anyone who needs to write and
+maintain multiple Dockerfiles or complex containerized environments that share
+interchangeable build steps and runtime requirements. This includes, but is not
+limited to, research groups and labs (e.g., robotics, vision, ML, compilers,
+systems), instructors who need reliable student environments, and continuous
+integration (CI) maintainers who prefer deterministic builds and centralized run
+policy over ad-hoc scripts.
+
+# Functionality
+
+`pythainer` is a lightweight Python package and CLI that provides a programmable
+front-end to Docker. It addresses the above pain points by adding a programmable
+abstraction for image construction and a reusable abstraction for execution
+policy. Builders are Python objects and functions that support ordinary
+programming constructs (conditionals, loops, parameters) and can be composed
+with a simple operator. Runners encapsulate repeatable `docker run` policy, so
+launching a container is a matter of selecting presets rather than rewriting
+long commands.
+
+Instead of writing raw Dockerfiles and shell scripts, users compose images with
+builders and control runtime behavior with runners. The library integrates
+naturally into Python workflows while emitting standard, human-readable
+Dockerfiles that are built and executed using the Docker engine with
+reproducible runtime settings.
+`pythainer` is centered around two core abstractions, builders and runners:
+
[…]
