### 2026-01-08 06a84e4
joss: Change bib file capitalization
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
