+++ doc/joss25/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{nvblox,
-  title        = {{nvblox: GPU-Accelerated Incremental Signed Distance Field Mapping}},
+  title        = {nvblox: {GPU}-Accelerated Incremental Signed Distance Field Mapping},
   author       = {Millane, Alexander and Oleynikova, Helen and Wirbel, Emilie and Steiner, Remo and Ramasamy, Vikram and Tingdahl, David and Siegwart, Roland},
   year         = 2024,
   booktitle    = {2024 IEEE International Conference on Robotics and Automation (ICRA)},
@@ -11,7 +11,7 @@
 }
 
 @inproceedings{robotcore,
-  title        = {{RobotCore: An Open Architecture for Hardware Acceleration in ROS 2}},
+  title        = {{RobotCore}: An Open Architecture for Hardware Acceleration in {ROS} 2},
   author       = {Mayoral-Vilches, Víctor and Neuman, Sabrina M. and Plancher, Brian and Reddi, Vijay Janapa},
   year         = 2022,
   booktitle    = {2022 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
@@ -23,7 +23,7 @@
 }
 
 @inproceedings{tani2020reproducible,
-  title        = {{Integrated Benchmarking and Design for Reproducible and Accessible Evaluation of Robotic Agents}},
+  title        = {Integrated Benchmarking and Design for Reproducible and Accessible Evaluation of Robotic Agents},
   author       = {Tani, Jacopo and Daniele, Andrea F. and Bernasconi, Gianmarco and Camus, Amaury and Petrov, Aleksandar and Courchesne, Anthony and Mehta, Bhairav and Suri, Rohit and Zaluska, Tomasz and Walter, Matthew R. and Frazzoli, Emilio and Paull, Liam and Censi, Andrea},
   year         = 2020,
   booktitle    = {2020 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
@@ -35,7 +35,7 @@
 }
 
 @article{macenski2022ros2,
-  title        = {{Robot Operating System 2: Design, architecture, and uses in the wild}},
+  title        = {Robot Operating System 2: Design, architecture, and uses in the wild},
   author       = {Steven Macenski  and Tully Foote  and Brian Gerkey  and Chris Lalancette  and William Woodall},
   year         = 2022,
   journal      = {Science Robotics},
@@ -48,7 +48,7 @@
 }
 
 @manual{nvidiaCudaPg13,
-  title        = {{CUDA C++ Programming Guide}},
+  title        = {{CUDA} {C++} Programming Guide},
   year         = 2025,
   note         = {Version 13.0},
   organization = {NVIDIA Corporation},
@@ -56,21 +56,21 @@
 }
 
 @misc{qemu,
-  title        = {{QEMU: A generic and open source machine emulator and virtualizer}},
-  author       = {{QEMU Project}},
+  title        = {{QEMU}: A generic and open source machine emulator and virtualizer},
+  author       = {{QEMU} Project},
   year         = 2003,
   howpublished = {\url{https://www.qemu.org/} accessed 2025-09-15}
 }
 
 @misc{docker,
-  title        = {{Docker: Accelerated Container Application Development}},
-  author       = {{Docker, Inc.}},
+  title        = {{Docker}: Accelerated Container Application Development},
+  author       = {{Docker}, Inc.},
   year         = 2013,
   howpublished = {\url{https://www.docker.com/} accessed 2025-09-15}
 }
 
 @inproceedings{llvm,
-  title        = {{LLVM: a compilation framework for lifelong program analysis & transformation}},
[…]
