### 2026-01-08 f338c04
joss: Revert changes in the bib
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
