+++ doc/joss25/paper.bib
@@ -57,14 +57,14 @@
 
 @misc{qemu,
   title        = {{QEMU}: A generic and open source machine emulator and virtualizer},
-  author       = {{QEMU} Project},
+  author       = {{QEMU Project}},
   year         = 2003,
   howpublished = {\url{https://www.qemu.org/} accessed 2025-09-15}
 }
 
 @misc{docker,
   title        = {{Docker}: Accelerated Container Application Development},
-  author       = {{Docker}, Inc.},
+  author       = {{Docker, Inc.}},
   year         = 2013,
   howpublished = {\url{https://www.docker.com/} accessed 2025-09-15}
 }
@@ -142,28 +142,28 @@
 
 @misc{docker_compose,
   title        = {{Docker} Compose},
-  author       = {{Docker}, Inc.},
+  author       = {{Docker, Inc.}},
   year         = 2014,
   howpublished = {\url{https://docs.docker.com/compose/} accessed 2025-09-15}
 }
 
 @misc{docker_sdk_python,
   title        = {{Docker} SDK for Python},
-  author       = {{Docker}, Inc.},
+  author       = {{Docker, Inc.}},
   year         = 2014,
   howpublished = {\url{https://docker-py.readthedocs.io/} accessed 2025-09-15}
 }
 
 @misc{devcontainers,
   title        = {Development Containers Specification},
-  author       = {Dev Containers Spec},
+  author       = {{Dev Containers Spec}},
   year         = 2022,
   howpublished = {\url{https://containers.dev} accessed 2025-09-15}
 }
 
 @misc{repo2docker,
   title        = {repo2docker: Turn repositories into {Jupyter}-enabled {Docker} images},
-  author       = {Project Jupyter},
+  author       = {{Project Jupyter}},
   year         = 2017,
   howpublished = {\url{https://repo2docker.readthedocs.io/} accessed 2025-09-15}
 }
