### 2026-01-08 #9059 @apaolillo
@danielskatz I just did revert here as per your commetns in the previous PR: https://github.com/apaolillo/pythainer/pull/54
