### 2026-01-08 3baa928
Introduce explicit build context for COPY (#49)
Replace ad-hoc filesystem staging with a first-class BuildContext that
tracks host paths and materializes a proper Docker build context.

Rework builder.copy() to register sources in the context, support
multi-source COPY, and forward --chown/--chmod flags explicitly. Make
CopyDockerBuildCommand pure by restricting it to context-relative paths
and Dockerfile string generation only.

Update the build flow to construct and use the generated context
directory, and clean up the COPY example accordingly.

Signed-off-by: Antonio Paolillo <apaolill@gmai
[…]
