### 2026-01-08 b459d6b
Release 0.0.6
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
