### 2026-01-08 77a9863
Release 0.0.6 (#55)
Signed-off-by: Antonio Paolillo <apaolill@gmail.com>
