### 2026-01-08 #9059 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
