### 2026-01-09 #9059 @danielskatz
Congratulations to @apaolillo (Antonio Paolillo) on your JOSS publication!!

And thanks to @samfrm, @akshaymittal143, and @rcannood for editing.
JOSS completely depends on volunteers, and we couldn't do this without you.
