### 2026-01-11 #9059 @samfrm
> Thanks [@danielskatz](https://github.com/danielskatz), [@samfrm](https://github.com/samfrm), [@akshaymittal143](https://github.com/akshaymittal143), and [@rcannood](https://github.com/rcannood)! This was a great experience, and I really appreciate the careful and constructive review process. I will definitely consider submitting to JOSS again in the future.

@apaolillo  Congratulations and good job 👏 🎉
